+++ pyproject.toml
@@ -16,7 +16,7 @@ classifiers = [
 ]
 
 dependencies = [
-    "jax>=0.4.31,<0.7.0",
+    "jax>=0.4.31",
     "tqdm>=4.66.4",
     "numpy>=1.24",
     "pandas",
@@ -25,7 +25,6 @@ dependencies = [
     "typing_extensions",
     "seaborn",
     "matplotlib",
-    "tensorflow-probability",
     "loess"
 ]
