+++ requirements.txt
@@ -1,4 +1,4 @@
-jax>=0.5.3,<0.7.0
+jax>=0.5.3
 tqdm>=4.66.4
 pypomp>=0.0.1
 numpy>=1.24
@@ -10,5 +10,4 @@ scipy
 typing_extensions
 seaborn
 matplotlib
-tensorflow-probability 
 loess
